+++ framework_src/native_library/android/FreKotlinExampleANE/FreKotlin/src/main/java/com/adobe/air/TRStateChangeCallback.java
@@ -15,5 +15,5 @@
  */
 
 package com.adobe.air;
-public abstract interface TRStateChangeCallback extends AndroidActivityWrapper.StateChangeCallback {
+public interface TRStateChangeCallback extends AndroidActivityWrapper.StateChangeCallback {
 }