+++ framework_src/native_library/android/FreKotlinExampleANE/FreKotlin/src/main/java/com/tuarua/frekotlin/FreKotlinContext.java
@@ -33,8 +33,9 @@ import java.util.HashMap;
 import java.util.Map;
 import java.lang.reflect.Method;
 //This file must remain as Java
+@SuppressWarnings("unused")
 public class FreKotlinContext extends FREContext implements TRActivityResultCallback, TRStateChangeCallback {
-    private String TAG = null;
+    private String TAG;
     protected FreKotlinMainController controller;
     private String[] functions;
     public FreKotlinContext(String name, FreKotlinMainController controller, String[] functions) {
@@ -70,6 +71,7 @@ public class FreKotlinContext extends FREContext implements TRActivityResultCall
     public void onConfigurationChanged(Configuration configuration) {
     }
 
+    @SuppressWarnings("JavaReflectionInvocation")
     private class CallKotlinFunction implements FREFunction {
         private String _name;
 
@@ -88,11 +90,11 @@ public class FreKotlinContext extends FREContext implements TRActivityResultCall
                 Method func = controller.getClass().getMethod(_name, parameterTypes);
                 return (FREObject) func.invoke(controller, freContext, al);
             } catch (NoSuchMethodException e) {
-                FREObject ret = new FreException(e, "can't call function " + _name).getError(e.getStackTrace());
+                FREObject ret = new FreException(e, "can't call function " + _name).getError();
                 Log.e(TAG, "can't find function " + _name + " ", e);
                 return ret;
             } catch (IllegalAccessException | InvocationTargetException e) {
-                FREObject ret = new FreException(e, "can't call function " + _name).getError(e.getStackTrace());
+                FREObject ret = new FreException(e, "can't call function " + _name).getError();
                 Log.e(TAG, "can't call function " + _name, e);
                 return ret;
             }