+++ starter_project/native_extension/ane/build-android.sh
@@ -1,34 +1,30 @@
 #!/bin/sh
 
 #Get the path to the script and trim to get the directory.
-echo "Setting path to current directory to:"
 pathtome=$0
 pathtome="${pathtome%/*}"
-AIR_SDK="/Users/User/sdks/AIR/AIRSDK_31"
+AIR_SDK="/Users/eoinlandy/SDKs/AIRSDK_32"
 
 PROJECTNAME=HelloWorldANE
 
-#Copy SWC into place.
-echo "Copying SWC into place."
+# Copy SWC into place.
 cp "$pathtome/../bin/$PROJECTNAME.swc" "$pathtome/"
 
-#Extract contents of SWC.
-echo "Extracting files form SWC."
+# Extract contents of SWC.
 unzip "$pathtome/$PROJECTNAME.swc" "library.swf" -d "$pathtome"
 
-#Copy library.swf to folders.
-echo "Copying library.swf into place."
+# Copy library.swf to folders.
 cp "$pathtome/library.swf" "$pathtome/platforms/android"
 
-echo "Copying Android aars into place"
+# Copying Android aars into place.
 cp "$pathtome/../../native_library/android/$PROJECTNAME/app/build/outputs/aar/app-release.aar" "$pathtome/platforms/android/app-release.aar"
-echo "getting Android jars"
+
+# Getting Android jars.
 unzip "$pathtome/platforms/android/app-release.aar" "classes.jar" -d "$pathtome/platforms/android"
 unzip "$pathtome/platforms/android/app-release.aar" "res/*" -d "$pathtome/platforms/android"
 mv "$pathtome/platforms/android/res" "$pathtome/platforms/android/com.mycompany.$PROJECTNAME-res"
 
 #Run the build command.
-echo "Building ANE."
 "$AIR_SDK"/bin/adt -package \
 -target ane "$pathtome/$PROJECTNAME.ane" "$pathtome/extension_android.xml" \
 -swc "$pathtome/$PROJECTNAME.swc" \
@@ -41,10 +37,10 @@ com.mycompany.$PROJECTNAME-res/. \
 com.mycompany.$PROJECTNAME-res/. \
 -platformoptions "$pathtome/platforms/android/platform.xml" \
 
+#Clean up.
 rm "$pathtome/platforms/android/classes.jar"
 rm "$pathtome/platforms/android/app-release.aar"
 rm "$pathtome/platforms/android/library.swf"
-
 rm "$pathtome/$PROJECTNAME.swc"
 rm "$pathtome/library.swf"
 rm -r "$pathtome/platforms/android/com.mycompany.$PROJECTNAME-res"