+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+### 1.6.0
+- Upgraded to Kotlin 1.3.11
+- Upgraded to AIR 32
+- Refactor starter projects AS3
+- Deprecate stackTraceElements param from `FreException.getError()`
+- Add `FREArray.insert()`
+- Add `FREArray.remove()`
+- Add `FREArray.isEmpty`
+- Small Refactor
+
 ### 1.5.0
 - Upgraded to Kotlin 1.2.61
 - Upgraded to AIR 31